Change SEGGER options to non blocking print.
In the previous config if you were printing a lot in your bootloader
debug and you didn't have an ICE connected, the boot would stall waiting
for the SEGGER to read the print buffer.  Better I think to just throw
away the oldest contents of the ring buffer if the ICE is not connected.